refactor: update @qlover/create-app with new build tools and dependencies
- Replaced Rollup with tsup for building the application, enhancing build performance and simplifying configuration.
- Updated package.json to reflect changes in dependencies and scripts, including the removal of outdated entries.
- Improved README with installation and usage instructions for better user guidance.
- Removed unnecessary files and streamlined the project structure for clarity.
