+++ packages/create-app/README.md
@@ -1,5 +1,17 @@
 ## @qlover/create-app
 
+快速创建和初始化新的应用程序的命令行工具。
+
+### 安装
+
+```bash
+npm install -g @qlover/create-app
 ```
+
+### 使用方法
+
+你可以通过以下方式使用：
+
+```bash
 npx @qlover/create-app
 ```
