--- packages/create-app/rollup.config.js
@@ -1,6 +1,0 @@
-import { createBaseRollup } from '../../make/createBaseRollup.js';
-import { Env } from '@qlover/env-loader';
-
-export default createBaseRollup({
-  isProduction: Env.searchEnv().get('NODE_ENV') === 'production'
-});
